split handy tools into files
