+++ lua/myplugins/handy-tools/snippets.lua
@@ -1,0 +1,6 @@
+return { 'SirVer/ultisnips', disable = false,
+    config = function()
+        vim.g.UltiSnipsEditSplit = 'tabdo'
+        vim.g.UltiSnipsSnippetDirectories = { vim.fn.stdpath('config') .. '/UltiSnips' }
+    end,
+}
