+++ lua/myplugins/handy-tools/treesitter-context.lua
@@ -1,0 +1,7 @@
+return { "nvim-treesitter/nvim-treesitter-context", disable = false,
+    requires = "nvim-treesitter/nvim-treesitter",
+    config = function()
+        require("treesitter-context").setup {}
+    end,
+}
+
