+++ lua/vimrc/plugins.lua
@@ -24,259 +24,18 @@ return require('packer').startup({function(use)
     --                                Handy Tools
     --~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-    use { 'thinca/vim-localrc', disable=true }
-    use { 'MunifTanjim/exrc.nvim', disable=false, -- local .nvimrc files {{{
-        requires = { 'MunifTanjim/nui.nvim', opt=true },
-        config = function()
-            -- disable built-in local config file support
-            vim.o.exrc = false
-            require("exrc").setup({
-                files = {
-                    ".nvimrc.lua",
-                    ".nvimrc",
-                    ".exrc.lua",
-                    ".exrc",
-                },
-            })
-        end,
-    } -- }}}
+    use(require('myplugins.handy-tools.local-vimrc'))
+    use(require('myplugins.handy-tools.multi-cursor'))
+    use(require('myplugins.handy-tools.printer'))
 
-    use { 'mg979/vim-visual-multi', branch = 'master', disable=false, -- {{{
-        config = function()
-            vim.g.VM_leader = [[\]]
-            vim.g.VM_theme = 'iceblue'
-            vim.g.VM_mouse_mappings = 1
-            -- mouse mappings not working? :(
-            vim.keymap.set('n', '<C-LeftMouse>', '<Plug>(VM-Mouse-Cursor)')
-            vim.keymap.set('n', '<C-RightMouse>', '<Plug>(VM-Mouse-Word)')
-            vim.keymap.set('n', '<M-C-RightMouse>', '<Plug>(VM-Mouse-Column)')
-            vim.g.VM_maps = {
-                -- https://github.com/mg979/vim-visual-multi/wiki/Quick-start#undoredo
-                ["Undo"] = 'u',
-                ["Redo"] = '<C-r>',
-            }
-        end,
-    } -- }}}
+    use(require('myplugins.handy-tools.telescope'))
+    use(require('myplugins.handy-tools.fzf-lua'))
+    use(require('myplugins.handy-tools.fzf-vim'))
 
-    use { 'nvim-telescope/telescope.nvim', disable=true, -- {{{
-        requires = {
-            {'nvim-lua/plenary.nvim'},
-            {'https://github.com/nvim-telescope/telescope-fzy-native.nvim'},
-            -- {'nvim-telescope/telescope-fzf-native.nvim', run = 'make' },
-        },
-        config = function()
-            local nmap = function(LH, RH, args) vim.keymap.set('n', LH, RH, args) end
-            local tb = require('telescope.builtin')
-            nmap('<leader>ff', function() tb.find_files() end)
-            nmap('<leader>fr', function() tb.live_grep() end)
-            nmap('<leader>fb', function() tb.buffers() end)
-            nmap('<leader>fl', function() tb.current_buffer_fuzzy_find() end)
-            nmap('<leader>fh', function() tb.help_tags() end)
-            nmap('<leader>fj', function() tb.jumplist() end)
-            nmap('<leader>fm', function() tb.marks() end)
-            nmap('<leader>f:', function() tb.command_history() end)
-            nmap('<leader>f/', function() tb.search_history() end)
-            nmap('<leader>ft', function() tb.treesitter() end)
-            -- nmap('<leader>ft', function() tb.filetypes() end)
-            nmap('z=', function() tb.spell_suggest() end)
+    use(require('myplugins.handy-tools.snippets'))
 
-            -- require('telescope.sorters').get_fzy_sorter()
-            -- https://github.com/nvim-telescope/telescope-fzy-native.nvim
-            require('telescope').setup {
-                extensions = {
-                    fzy_native = {
-                        override_generic_sorter = true,
-                        override_file_sorter = true,
-                    }
-                }
-            }
-            require('telescope').load_extension('fzy_native')
-        end,
-    } -- }}}
-
-    use { 'junegunn/fzf', run = './install --xdg --all', disable=false }
-    use { 'junegunn/fzf.vim', disable=true, -- {{{
-        config = function()
-            vim.g.fzf_preview_window = { 'right,50%,<70(up,40%)', 'alt-p' }
-
-            local nmap = function(LH, RH, args) vim.keymap.set('n', LH, RH, args) end
-            local imap = function(LH, RH, args) vim.keymap.set('i', LH, RH, args) end
-
-            -- FZF Commands
-            nmap('<Leader>fr', ':Rg<CR>')
-            nmap('<Leader>ff', ':Files<CR>')
-            nmap('<Leader>fg', ':GFiles<CR>')
-            nmap('<Leader>fb', ':Buffers<CR>')
-            nmap('<Leader>fl', ':BLines<CR>')
-            nmap('<Leader>fL', ':Lines<CR>')
-            nmap('<Leader>fh', ':Helptags<CR>')
-            nmap('<Leader>fm', ':Marks<CR>')
-            nmap('<Leader>f:', ':History:<CR>')
-            nmap('<Leader>f/', ':History/:<CR>')
-            nmap('<Leader>fs', ':Snippets:<CR>')
-
-            -- Insert mode completion
-            imap('<Leader><c-x><c-k>', '<plug>(fzf-complete-word)')
-            imap('<Leader><c-x><c-f>', '<plug>(fzf-complete-path)')
-            imap('<Leader><c-x><c-j>', '<plug>(fzf-complete-file-ag)')
-            imap('<Leader><c-x><c-l>', '<plug>(fzf-complete-line)')
-            imap('<Leader><c-x><c-L>', '<plug>(fzf-complete-buffer-line)')
-        end,
-    } -- }}}
-    use { 'ibhagwan/fzf-lua', disable=false,
-        requires = {
-            -- 'nvim-tree/nvim-web-devicons',
-            { 'junegunn/fzf', run = './install --xdg --all' },
-        },
-        config = function()
-            local fzflua = require('fzf-lua')
-            fzflua.setup{
-                keymap = { --  {{{
-                    builtin = {
-                        -- neovim `:tmap` mappings for the fzf win
-                        ["<F1>"]  = "toggle-help",
-                        ["<F2>"]  = "toggle-fullscreen",
-                        -- Only valid with the 'builtin' previewer
-                        ["<F3>"]  = "toggle-preview-wrap",
-                        ["<F4>"]  = "toggle-preview",
-                        -- Rotate preview clockwise/counter-clockwise
-                        ["<F5>"]  = "toggle-preview-ccw",
-                        ["<F6>"]  = "toggle-preview-cw",
-                        ["<c-d>"] = "preview-page-down",
-                        ["<c-u>"] = "preview-page-up",
-                        ["<c-h>"] = "preview-page-reset",
-                    },
-                    fzf = {
-                        -- fzf '--bind=' options
-                        -- ["ctrl-u"] = "unix-line-discard",
-                        ["ctrl-f"] = "half-page-down",
-                        ["ctrl-b"] = "half-page-up",
-                        ["ctrl-a"] = "beginning-of-line",
-                        ["ctrl-e"] = "end-of-line",
-                        ["alt-a"]  = "toggle-all",
-                        -- Only valid with fzf previewers (bat/cat/git/etc)
-                        ["f3"]     = "toggle-preview-wrap",
-                        ["f4"]     = "toggle-preview",
-                        ["ctrl-d"] = "preview-page-down",
-                        ["ctrl-u"] = "preview-page-up",
-                    },
-                }, --  }}}
-                fzf_opts = {
-                    ['--layout'] = 'default',
-                }
-            }
-
-            local nmap = function(LH, RH, args) vim.keymap.set('n', LH, RH, args) end
-            local imap = function(LH, RH, args) vim.keymap.set('i', LH, RH, args) end
-
-            -- files
-            -- nmap('<Leader>fr', function() fzflua.grep_project() end)
-            nmap('<Leader>fr', function() fzflua.live_grep_native({ exec_empty_query = true }) end)
-            nmap('<Leader>ff', function() fzflua.files() end)
-            nmap('<Leader>fg', function() fzflua.git_files() end)
-            nmap('<Leader>fb', function() fzflua.buffers() end)
-            nmap('<Leader>fl', function() fzflua.blines() end)
-            nmap('<Leader>fL', function() fzflua.lines() end)
-
-            -- lists
-            nmap('<Leader>fh', function() fzflua.help_tags() end)
-            nmap('<Leader>fj', function() fzflua.jumps() end)
-            nmap('<Leader>fc', function() fzflua.changes() end)
-            nmap('<Leader>fm', function() fzflua.marks() end)
-            nmap('<Leader>fR', function() fzflua.registers() end)
-            nmap('<Leader>f:', function() fzflua.command_history() end)
-            nmap('<Leader>f/', function() fzflua.search_history() end)
-            nmap('<Leader>fs', function() print('hello') end)
-
-            -- suggestions
-            nmap('z=', function() fzflua.spell_suggest() end)
-
-            -- Insert mode completion
-            imap('<Leader><c-x><c-k>', function() print('hello') end)
-            imap('<Leader><c-x><c-f>', function() print('hello') end)
-            imap('<Leader><c-x><c-j>', function() print('hello') end)
-            imap('<Leader><c-x><c-l>', function() print('hello') end)
-            imap('<Leader><c-x><c-L>', function() print('hello') end)
-
-            -- `:help vim.ui.select` for more info
-            fzflua.register_ui_select()
-        end,
-    }
-
-    use { 'SirVer/ultisnips', disable=false, -- {{{
-        config = function()
-            vim.g.UltiSnipsEditSplit = 'tabdo'
-            vim.g.UltiSnipsSnippetDirectories = { vim.fn.stdpath('config') .. '/UltiSnips' }
-        end,
-    } -- }}}
-
-    use { 'rareitems/printer.nvim', disable = false, --  {{{
-        config = function()
-            require('printer').setup({
-                keymap = "gp",
-                -- add_to_inside = function(text)
-                --     return text
-                -- end,
-                formatters = {
-                    python = function(inside, variable)
-                        return string.format([[print(f'%s = {%s}')]], inside, variable)
-                    end
-                },
-            })
-        end
-    } --  }}}
-
-    use { 'lewis6991/gitsigns.nvim', disable=false, --  {{{
-        config = function()
-            require('gitsigns').setup{
-                on_attach = function(bufnr)
-                    local gs = package.loaded.gitsigns
-
-                    local function map(mode, l, r, opts)
-                        opts = opts or {}
-                        opts.buffer = bufnr
-                        vim.keymap.set(mode, l, r, opts)
-                    end
-
-                    -- Navigation
-                    map('n', ']c', function()
-                        if vim.wo.diff then return ']c' end
-                        vim.schedule(function() gs.next_hunk() end)
-                        return '<Ignore>'
-                    end, {expr=true})
-
-                    map('n', '[c', function()
-                        if vim.wo.diff then return '[c' end
-                        vim.schedule(function() gs.prev_hunk() end)
-                        return '<Ignore>'
-                    end, {expr=true})
-
-                    -- Actions
-                    map({'n', 'v'}, '<leader>ghs', ':Gitsigns stage_hunk<CR>')
-                    map({'n', 'v'}, '<leader>ghr', ':Gitsigns reset_hunk<CR>')
-                    map('n', '<leader>ghS', gs.stage_buffer, { desc = 'Gitsigns: Stage buffer' })
-                    map('n', '<leader>ghR', gs.reset_buffer, { desc = 'Gitsigns: Reset buffer' })
-                    map('n', '<leader>ghu', gs.undo_stage_hunk, { desc = 'Gitsigns: undo stage hunk' })
-                    map('n', '<leader>ghp', gs.preview_hunk, { desc = 'Gitsigns: preview hunk' })
-                    map('n', '<leader>ghb', function() gs.blame_line{full=true} end, { desc = 'Gitsigns: Blame line' })
-                    map('n', '<leader>gtb', gs.toggle_current_line_blame, { desc = 'Gitsigns: toggle current line blame' })
-                    map('n', '<leader>ghd', gs.diffthis, { desc = 'Gitsigns: diff this' })
-                    map('n', '<leader>ghD', function() gs.diffthis('~') end, { desc = 'Gitsigns: diff this ~' })
-                    map('n', '<leader>gtd', gs.toggle_deleted, { desc = 'Gitsigns: toggle showing deleted lines' })
-
-                    -- Text object
-                    map({'o', 'x'}, 'gih', ':<C-U>Gitsigns select_hunk<CR>')
-                end
-            }
-        end
-    } --  }}}
-
-    use { "nvim-treesitter/nvim-treesitter-context", disable=false, --  {{{
-        requires = "nvim-treesitter/nvim-treesitter",
-        config = function()
-            require("treesitter-context").setup{}
-        end,
-    } --  }}}
+    use(require('myplugins.handy-tools.gitsigns'))
+    use(require('myplugins.handy-tools.treesitter-context'))
 
     --~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     --                                  Writing
